Stop image blocks from growing bigger than their column
If the user selects a large image then it could end up displaying larger
than the space given to the block by the user. It looks a bit gross, but
in order to have images stay within their bounds they will be squashed
to fit. In the future we can probably calculate a more appropriate
value for max-width.